+++ x-pack/test/functional_with_es_ssl/apps/cases/group1/create_case_form.ts
@@ -8,12 +8,12 @@
 import expect from '@kbn/expect';
 import uuid from 'uuid';
 import { CaseSeverity } from '@kbn/cases-plugin/common/api';
-import { FtrProviderContext } from '../../ftr_provider_context';
+import { FtrProviderContext } from '../../../ftr_provider_context';
 import {
   createUsersAndRoles,
   deleteUsersAndRoles,
-} from '../../../cases_api_integration/common/lib/authentication';
-import { users, roles, casesAllUser, casesAllUser2 } from './common';
+} from '../../../../cases_api_integration/common/lib/authentication';
+import { users, roles, casesAllUser, casesAllUser2 } from '../common';
 
 export default ({ getService, getPageObject }: FtrProviderContext) => {
   describe('Create case', function () {
