+++ x-pack/test/functional_with_es_ssl/apps/cases/group1/index.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { FtrProviderContext } from '../../../ftr_provider_context';
+
+export default ({ loadTestFile }: FtrProviderContext) => {
+  describe('Cases - group 1', function () {
+    loadTestFile(require.resolve('./create_case_form'));
+    loadTestFile(require.resolve('./view_case'));
+    loadTestFile(require.resolve('./deletion'));
+  });
+};
