+++ x-pack/test/functional_with_es_ssl/apps/discover_ml_uptime/ml/index.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { FtrProviderContext } from '../../ftr_provider_context';
+import { FtrProviderContext } from '../../../ftr_provider_context';
 
 export default ({ loadTestFile, getService }: FtrProviderContext) => {
   const ml = getService('ml');
