+++ x-pack/test/functional_with_es_ssl/apps/discover_ml_uptime/uptime/alert_flyout.ts
@@ -8,7 +8,7 @@
 import expect from '@kbn/expect';
 import { setTimeout as setTimeoutAsync } from 'timers/promises';
 
-import { FtrProviderContext } from '../../ftr_provider_context';
+import { FtrProviderContext } from '../../../ftr_provider_context';
 
 export default ({ getPageObjects, getService }: FtrProviderContext) => {
   describe('uptime alerts', () => {
