+++ x-pack/test/functional_with_es_ssl/apps/discover_ml_uptime/uptime/simple_down_alert.ts
@@ -6,8 +6,8 @@
  */
 
 import expect from '@kbn/expect';
-import { FtrProviderContext } from '../../ftr_provider_context';
-import { deleteUptimeSettingsObject } from '../../../functional/apps/uptime';
+import { FtrProviderContext } from '../../../ftr_provider_context';
+import { deleteUptimeSettingsObject } from '../../../../functional/apps/uptime';
 
 export default ({ getPageObjects, getService }: FtrProviderContext) => {
   describe('uptime simple status alert', () => {
